Added item counter at the top of the page
